feat: Zarr dataflow implementation Step 4-6  (Add Python ingestion stage: shard → Zarr AND Update incremental_train.py to train from Zarr, not CSV AND Update dependency bootstrap)

+++ simulation_and_vr/imitationagent_training/training_pipeline/data_contract.py
@@ -32,7 +32,7 @@ TIMESTAMP_SHAPE = ("time",)
 
 # Storage types are explicit to avoid implicit conversion bugs during training.
 RGB_DTYPE = "uint8"
-DEPTH_DTYPE = "float16"
+DEPTH_DTYPE = "float32"
 VECTOR_DTYPE = "float32"
 TIMESTAMP_DTYPE = "float64"
